+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/scheduling/resourcePolicies/ExecutionClusterInfo.scala
@@ -1,0 +1,13 @@
+package edu.uci.ics.amber.engine.architecture.scheduling.resourcePolicies
+
+import edu.uci.ics.amber.engine.architecture.scheduling.Region
+
+object ExecutionClusterInfo {
+  def apply(): ExecutionClusterInfo = {
+    ExecutionClusterInfo()
+  }
+}
+
+case class ExecutionClusterInfo() {
+
+}
